+++ crates/core/config/src/toolchain/mod.rs
@@ -1,9 +1,11 @@
 mod config;
 mod deno;
 mod node;
+mod rust;
 mod typescript;
 
 pub use config::*;
 pub use deno::*;
 pub use node::*;
+pub use rust::*;
 pub use typescript::*;