+++ crates/rust/lang/src/lib.rs
@@ -1,10 +1,13 @@
+pub mod cargo_lock;
+pub mod cargo_toml;
+
 use moon_lang::{DependencyManager, Language, VersionManager};
 
 pub const RUST: Language = Language {
     binary: "rustc",
     file_exts: &["rs", "rlib"],
     vendor_bins_dir: None,
-    vendor_dir: None,
+    vendor_dir: Some("target"),
 };
 
 // Package managers
@@ -21,4 +24,9 @@ pub const CARGO: DependencyManager = DependencyManager {
 pub const RUSTUP: VersionManager = VersionManager {
     binary: "rustup",
     version_file: "rust-toolchain.toml", // A config file
+};
+
+pub const RUSTUP_LEGACY: VersionManager = VersionManager {
+    binary: "rustup",
+    version_file: "rust-toolchain", // A config file
 };