+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Added Rust tier 2 language support!
+  - Added `rust` setting to `.moon/toolchain.yml`.
+  - Added `rust` as a supported `platform` variant.
+
 ## 1.4.0
 
 #### 🚀 Updates